+++ lib/active_record/userstamp/controller_additions.rb
@@ -8,13 +8,21 @@ module ActiveRecord::Userstamp::ControllerAdditions
   extend ActiveSupport::Concern
 
   included do
-    before_filter :set_stamper
-    after_filter :reset_stamper
+    around_action :with_stamper
   end
 
   private
 
-  # The +set_stamper+ method as implemented here assumes a couple of things. First, that you are
+  # This {#with_stamper} method sets the stamper for the duration of the action. This ensures
+  # that exceptions raised within the controller action would properly restore the previous stamper.
+  def with_stamper
+    set_stamper
+    yield
+  ensure
+    reset_stamper
+  end
+
+  # The {#set_stamper} method as implemented here assumes a couple of things. First, that you are
   # using a +User+ model as the stamper and second that your controller has a +current_user+
   # method that contains the currently logged in stamper. If either of these are not the case in
   # your application you will want to manually add your own implementation of this method to the
@@ -23,7 +31,7 @@ module ActiveRecord::Userstamp::ControllerAdditions
     ActiveRecord::Userstamp.config.default_stamper_class.push_stamper(current_user)
   end
 
-  # The +reset_stamper+ method as implemented here assumes that a +User+ model is being used as
+  # The {#reset_stamper} method as implemented here assumes that a +User+ model is being used as
   # the stamper. If this is not the case then you will need to manually add your own
   # implementation of this method to the private section of your +ApplicationController+
   def reset_stamper